requirements.txt
# langchain==0.2.7
# langchain_community==0.2.7
# langchain_chroma==0.1.2
# PyPDF2==3.0.1
# streamlit==1.36.0
# pandas==2.2.2

langchain==0.1.20
langchain_community==0.0.38
pandas==1.5.3
docx2txt==0.8
pandasai==2.2.10
python-dotenv==1.0.1
streamlit==1.36.0
PyPDF2==3.0.1